refactor(fleet): extract capabilityGrouping to a pure module (CodeRabbit)
Move groupCapabilities + CapabilityGroup + GROUP_RENDER_ORDER out of the
PermissionApprovalDialog component into a JSX-free capabilityGrouping module so
the approval-inbox selector no longer imports a .tsx (store/selector → UI
component layering inversion). The dialog re-exports both symbols for backward
compat; ApprovalInboxList and the dialog test are unchanged. Also drop an unused
evaluate-callback param in the dogfood harness. No behavior change.
