+++ src/renderer/components/Approval/PermissionApprovalDialog.tsx
@@ -15,13 +15,14 @@
 // visible — "can label your panes" sits in neutral grey, "can read what's
 // on your screen" sits in warning red.
 
-import {
-  CAPABILITY_RISK_CLASS,
-  RISK_CLASS_COPY,
-  type RiskClass,
-  type RiskClassCopy,
-} from '../../../main/mcp/methodCapabilityMap';
-import { parsePermission } from '../../../main/mcp/permissionGrammar';
+import { type RiskClassCopy } from '../../../main/mcp/methodCapabilityMap';
+import { groupCapabilities } from './capabilityGrouping';
+
+// Re-export the grouping helpers so existing importers of these symbols from
+// the dialog module keep working — the canonical home is now the pure
+// `./capabilityGrouping` module (decouples non-UI consumers from this `.tsx`).
+export { groupCapabilities } from './capabilityGrouping';
+export type { CapabilityGroup } from './capabilityGrouping';
 
 export interface PermissionApprovalDialogProps {
   /** Plugin name (e.g. `claude-ai`, `my-org.my-tool`). */
@@ -39,63 +40,6 @@ export interface PermissionApprovalDialogProps {
   onApprove: () => void;
   /** Called when the user clicks Deny. */
   onDeny: () => void;
-}
-
-/**
- * One row in the dialog's grouped capability list. Exported so tests can
- * verify the grouping logic separately from the render.
- */
-export interface CapabilityGroup {
-  riskClass: RiskClass;
-  copy: RiskClassCopy;
-  capabilities: { raw: string; capability: string; pathGlob?: string }[];
-}
-
-/**
- * Group declared capabilities by risk class. Order of groups follows
- * `GROUP_RENDER_ORDER` so critical items always appear first regardless of
- * declaration order in the wire payload.
- */
-const GROUP_RENDER_ORDER: RiskClass[] = [
-  'terminal-content',
-  'terminal-input',
-  'browser',
-  'a2a',
-  'metadata',
-  'events',
-  'pane-lifecycle',
-  'workspace',
-  'internal',
-];
-
-export function groupCapabilities(
-  declared: readonly string[],
-): CapabilityGroup[] {
-  const groups = new Map<RiskClass, CapabilityGroup>();
-  for (const raw of declared) {
-    const parsed = parsePermission(raw);
-    if (!parsed.ok) continue; // skip malformed; substrate already rejected at declare-time
-    const cap = parsed.permission.capability;
-    const risk =
-      CAPABILITY_RISK_CLASS[cap] ?? ('metadata' as RiskClass); // safe fallback
-    let group = groups.get(risk);
-    if (!group) {
-      group = {
-        riskClass: risk,
-        copy: RISK_CLASS_COPY[risk],
-        capabilities: [],
-      };
-      groups.set(risk, group);
-    }
-    group.capabilities.push({
-      raw,
-      capability: cap,
-      pathGlob: parsed.permission.pathGlob,
-    });
-  }
-  return GROUP_RENDER_ORDER.filter((r) => groups.has(r)).map(
-    (r) => groups.get(r) as CapabilityGroup,
-  );
 }
 
 function severityAccent(severity: RiskClassCopy['severity']): string {
